Implement SSR for LayoutManager
