+++ testing/tests/DevExpress.serverSide/formLayoutManager.tests.js
@@ -1,0 +1,3 @@
+"use strict";
+
+require("../DevExpress.ui.widgets.form/formLayoutManager.markup.tests.js");
